requirements.txt
discord.py[voice]==2.3.2
youtube_dl
